pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel", "setuptools-git-versioning<2"]
build-backend = "setuptools.build_meta"

[project]
name = "dafni-cli"
description = "A CLI to access the DAFNI APIs from the command line"
readme = "README.md"
authors = [{ name = "DAFNI Facility", email = "support@dafni.ac.uk" }]
dependencies = [
    "click>=8.1.6",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "tabulate>=0.9.0",
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
# Dynamic versioning using setuptools-git-versioning
dynamic = ["version"]

[tool.setuptools-git-versioning]
enabled = true

# To override the versioning above, comment out and replace with
# version = "0.0.1"

[project.urls]
Source = "https://github.com/dafnifacility/cli"

[project.scripts]
dafni = "dafni_cli.dafni:dafni"

[options]
package_dir = "dafni_cli"
